serve/FlaskApp/requirements.txt
flask==1.1.2
flask_bootstrap==3.3.7
requests==2.25.0
numpy==1.19.4
pillow==8.1.0
tensorflow==2.4.0
pandas==0.22.0

